requirements.txt
openai-whisper
torch
openai>=1.30.0
